Add logout and polling
