+++ src/components/SignUp/SignUpFooter.tsx
@@ -51,17 +51,7 @@ const SignUpFooter: React.FC<{
   const continueText =
     step === ('privacy' as any) ? 'Accept & Continue' : 'Continue';
 
-  const setPrivacyAgree = () => {
-    const target = {
-      name: 'privacyAgreement',
-      value: true
-    };
-
-    setForm({ target });
-  };
-
-  const privacyAgree = async () => {
-    await setPrivacyAgree();
+  const privacyAgree = () => {
     submitBasics();
     next();
   };
