+++ src/pages/App.tsx
@@ -1,4 +1,4 @@
-import React, { useEffect, useState } from 'react';
+import React from 'react';
 import {
   Redirect,
   Route,
@@ -16,41 +16,32 @@ import SignUp from './SignUp';
 import Header from '../components/layout/Header';
 import Footer from '../components/layout/Footer';
 import ScrollToTop from '../components/ScrollToTop';
-import { getSessionCookie } from '../utils/session';
 import '../styles/App.scss';
 import GlobalStyle from '../theme/globalStyles';
 
-const CAG = () => {
-  const [session, setSession] = useState(getSessionCookie());
-
-  useEffect(() => {
-    setSession(getSessionCookie());
-  }, []);
-
-  return (
-    <Router>
-      <ScrollToTop />
-      <GlobalStyle />
-      <Header showSession={!!(session.userId !== undefined && session.active_status)} />
-      <Switch>
-        <Route
-          exact
-          path="/"
-          render={() => <Redirect to={{ pathname: '/home' }} />}
-        />
-        <Route component={Home} exact path="/home" />
-        <Route component={FAQ} exact path="/faq" />
-        <Route component={TOS} exact path="/terms-of-service" />
-        <Route component={WhoWeAre} exact path="/who-we-are" />
-        <Route component={Login} exact path="/login" />
-        <Route component={Logout} exact path="/logout" />
-        <Route component={SignUp} exact path="/sign-up" />
-        <Route component={NotFound} />
-      </Switch>
-      <Footer />
-    </Router>
-  );
-};
+const CAG = () => (
+  <Router>
+    <ScrollToTop />
+    <GlobalStyle />
+    <Header />
+    <Switch>
+      <Route
+        exact
+        path="/"
+        render={() => <Redirect to={{ pathname: '/home' }} />}
+      />
+      <Route component={Home} exact path="/home" />
+      <Route component={FAQ} exact path="/faq" />
+      <Route component={TOS} exact path="/terms-of-service" />
+      <Route component={WhoWeAre} exact path="/who-we-are" />
+      <Route component={Login} exact path="/login" />
+      <Route component={Logout} exact path="/logout" />
+      <Route component={SignUp} exact path="/sign-up" />
+      <Route component={NotFound} />
+    </Switch>
+    <Footer />
+  </Router>
+);
 
 const App = () => (
   <main id="cag-frontend-app">
