+++ src/pages/Logout.tsx
@@ -1,5 +1,5 @@
 import React, { useEffect } from 'react';
-import { createBrowserHistory } from 'history';
+import { Redirect } from 'react-router-dom';
 import Cookies from 'js-cookie';
 import Row from 'react-bootstrap/Row';
 import Col from 'react-bootstrap/Col';
@@ -7,21 +7,16 @@ import PageContainer from '../components/layout/PageContainer';
 import { Title } from '../components/layout/Titles';
 
 const Logout = () => {
-  const history = createBrowserHistory();
-
-  useEffect(
-    () => {
-      Cookies.remove('session');
-      history.push('/login');
-    },
-    [history]
-  );
+  useEffect(() => {
+    Cookies.remove('session');
+  }, []);
 
   return (
     <PageContainer>
       <Row>
         <Col lg="8">
           <Title>Logging out...</Title>
+          <Redirect to="/login" />
         </Col>
       </Row>
     </PageContainer>
